+++ Changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  * Error when the `abi3` feature is selected but no minimum version
  * Support building universal2 wheels (x86 and aarch64 in a single file) by messense in [#403](https://github.com/PyO3/maturin/pull/403)
  * Recognize `PYO3_CROSS_LIB_DIR` for cross compiling with abi3 targeting windows. 
+ * `package.metadata.maturin.classifier` is renamed to `classifiers`
 
 ## 0.9.0 - 2020-01-10